+++ src/PhoneNumber.js
@@ -2,9 +2,7 @@ import { GraphQLScalarType } from 'graphql';
 import { GraphQLError } from 'graphql/error';
 import { Kind } from 'graphql/language';
 
-const PHONE_NUMBER_REGEX = new RegExp(
-  /^\+\d{11,15}$/,
-);
+const PHONE_NUMBER_REGEX = new RegExp(/^\+\d{7,15}$/);
 
 export default new GraphQLScalarType({
   name: 'PhoneNumber',
@@ -18,7 +16,9 @@ export default new GraphQLScalarType({
     }
 
     if (!PHONE_NUMBER_REGEX.test(value)) {
-      throw new TypeError(`Value is not a valid phone number of the form +17895551234 (10-15 digits): ${value}`);
+      throw new TypeError(
+        `Value is not a valid phone number of the form +17895551234 (7-15 digits): ${value}`,
+      );
     }
 
     return value;
@@ -30,7 +30,9 @@ export default new GraphQLScalarType({
     }
 
     if (!PHONE_NUMBER_REGEX.test(value)) {
-      throw new TypeError(`Value is not a valid phone number of the form +17895551234 (10-15 digits): ${value}`);
+      throw new TypeError(
+        `Value is not a valid phone number of the form +17895551234 (7-15 digits): ${value}`,
+      );
     }
 
     return value;
@@ -44,7 +46,9 @@ export default new GraphQLScalarType({
     }
 
     if (!PHONE_NUMBER_REGEX.test(ast.value)) {
-      throw new TypeError(`Value is not a valid phone number of the form +17895551234 (10-15 digits): ${ast.value}`);
+      throw new TypeError(
+        `Value is not a valid phone number of the form +17895551234 (7-15 digits): ${ast.value}`,
+      );
     }
 
     return ast.value;