+++ src/__tests__/PhoneNumber.test.js
@@ -24,21 +24,30 @@ describe('PhoneNumber', () => {
   describe('invalid', () => {
     describe('not a phone number', () => {
       test('serialize', () => {
-        expect(() => PhoneNumber.serialize('this is not a phone number')).toThrow(
-          /^Value is not a valid phone number of the form \+17895551234 \(10-15 digits\)/,
+        expect(() =>
+          PhoneNumber.serialize('this is not a phone number'),
+        ).toThrow(
+          /^Value is not a valid phone number of the form \+17895551234 \(7-15 digits\)/,
         );
       });
 
       test('parseValue', () => {
-        expect(() => PhoneNumber.parseValue('this is not a phone number')).toThrow(
-          /^Value is not a valid phone number of the form \+17895551234 \(10-15 digits\)/,
+        expect(() =>
+          PhoneNumber.parseValue('this is not a phone number'),
+        ).toThrow(
+          /^Value is not a valid phone number of the form \+17895551234 \(7-15 digits\)/,
         );
       });
 
       test('parseLiteral', () => {
         expect(() =>
-          PhoneNumber.parseLiteral({ value: 'this is not a phone number', kind: Kind.STRING }),
-        ).toThrow(/^Value is not a valid phone number of the form \+17895551234 \(10-15 digits\)/);
+          PhoneNumber.parseLiteral({
+            value: 'this is not a phone number',
+            kind: Kind.STRING,
+          }),
+        ).toThrow(
+          /^Value is not a valid phone number of the form \+17895551234 \(7-15 digits\)/,
+        );
       });
     });
 
@@ -48,73 +57,84 @@ describe('PhoneNumber', () => {
       });
 
       test('parseValue', () => {
-        expect(() => PhoneNumber.parseValue(123)).toThrow(/Value is not string/);
-      });
-
-      test('parseLiteral', () => {
-        expect(() => PhoneNumber.parseLiteral({ value: 123, kind: Kind.INT })).toThrow(
-          /Can only validate strings as phone numbers but got a/,
-        );
-      });
-    });
-
-    describe('too short', () => {
-      test('serialize', () => {
-        expect(() => PhoneNumber.serialize('+1789555123')).toThrow(
-          /^Value is not a valid phone number of the form \+17895551234 \(10-15 digits\)/,
-        );
-      });
-
-      test('parseValue', () => {
-        expect(() => PhoneNumber.parseValue('+1789555123')).toThrow(
-          /^Value is not a valid phone number of the form \+17895551234 \(10-15 digits\)/,
+        expect(() => PhoneNumber.parseValue(123)).toThrow(
+          /Value is not string/,
         );
       });
 
       test('parseLiteral', () => {
         expect(() =>
-          PhoneNumber.parseLiteral({ value: '+1789555123', kind: Kind.STRING }),
-        ).toThrow(/^Value is not a valid phone number of the form \+17895551234 \(10-15 digits\)/);
+          PhoneNumber.parseLiteral({ value: 123, kind: Kind.INT }),
+        ).toThrow(/Can only validate strings as phone numbers but got a/);
+      });
+    });
+
+    describe('too short', () => {
+      test('serialize', () => {
+        expect(() => PhoneNumber.serialize('+178 123')).toThrow(
+          /^Value is not a valid phone number of the form \+17895551234 \(7-15 digits\)/,
+        );
+      });
+
+      test('parseValue', () => {
+        expect(() => PhoneNumber.parseValue('+178 123')).toThrow(
+          /^Value is not a valid phone number of the form \+17895551234 \(7-15 digits\)/,
+        );
+      });
+
+      test('parseLiteral', () => {
+        expect(() =>
+          PhoneNumber.parseLiteral({ value: '+178 123', kind: Kind.STRING }),
+        ).toThrow(
+          /^Value is not a valid phone number of the form \+17895551234 \(7-15 digits\)/,
+        );
       });
     });
 
     describe('too long', () => {
       test('serialize', () => {
         expect(() => PhoneNumber.serialize('+1789555123456789')).toThrow(
-          /^Value is not a valid phone number of the form \+17895551234 \(10-15 digits\)/,
+          /^Value is not a valid phone number of the form \+17895551234 \(7-15 digits\)/,
         );
       });
 
       test('parseValue', () => {
         expect(() => PhoneNumber.parseValue('+1789555123456789')).toThrow(
-          /^Value is not a valid phone number of the form \+17895551234 \(10-15 digits\)/,
+          /^Value is not a valid phone number of the form \+17895551234 \(7-15 digits\)/,
         );
       });
 
       test('parseLiteral', () => {
         expect(() =>
-          PhoneNumber.parseLiteral({ value: '+1789555123456789', kind: Kind.STRING }),
-        ).toThrow(/^Value is not a valid phone number of the form \+17895551234 \(10-15 digits\)/);
+          PhoneNumber.parseLiteral({
+            value: '+1789555123456789',
+            kind: Kind.STRING,
+          }),
+        ).toThrow(
+          /^Value is not a valid phone number of the form \+17895551234 \(7-15 digits\)/,
+        );
       });
     });
 
     describe('no plus sign', () => {
       test('serialize', () => {
         expect(() => PhoneNumber.serialize('17895551234')).toThrow(
-          /^Value is not a valid phone number of the form \+17895551234 \(10-15 digits\)/,
+          /^Value is not a valid phone number of the form \+17895551234 \(7-15 digits\)/,
         );
       });
 
       test('parseValue', () => {
         expect(() => PhoneNumber.parseValue('17895551234')).toThrow(
-          /^Value is not a valid phone number of the form \+17895551234 \(10-15 digits\)/,
+          /^Value is not a valid phone number of the form \+17895551234 \(7-15 digits\)/,
         );
       });
 
       test('parseLiteral', () => {
         expect(() =>
           PhoneNumber.parseLiteral({ value: '17895551234', kind: Kind.STRING }),
-        ).toThrow(/^Value is not a valid phone number of the form \+17895551234 \(10-15 digits\)/);
+        ).toThrow(
+          /^Value is not a valid phone number of the form \+17895551234 \(7-15 digits\)/,
+        );
       });
     });
   });